FriendshipService/requirements.txt
fastapi==0.90.0
sqlalchemy==2.0.0
sqlalchemy[asyncio]==2.0.0
asyncmy==0.2.10
pytest==8.2.0
pytest-asyncio==0.19.0
httpx==0.23.0  # Si estás haciendo pruebas HTTP con TestClient de FastAPI
uvicorn==0.18.0  # Si estás ejecutando FastAPI con Uvicorn
pymysql==1.1.1
pydantic==1.10.13
python-dotenv==1.0.1
